Update for Educational Resource controller
