+++ backend/ims_application/src/main/java/com/swp391/ims_application/entity/EducationalResource.java
@@ -1,6 +1,5 @@
 package com.swp391.ims_application.entity;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
@@ -31,7 +30,6 @@ public class EducationalResource {
     @Column(name = "created_date")
     private Date createdDate;
 
-    @JsonProperty("isAvailable")
     @Column(name = "is_available")
     private boolean isAvailable;
