+++ backend/ims_application/src/main/java/com/swp391/ims_application/repository/EducationalResourceRepository.java
@@ -11,5 +11,6 @@ import java.util.List;
 public interface EducationalResourceRepository extends JpaRepository<EducationalResource, Integer> {
     List<EducationalResource> findAll();
 
+    EducationalResource findByResourceId(int id);
 }
