+++ backend/ims_application/src/main/java/com/swp391/ims_application/repository/ProgramTrainingResourceRepository.java
@@ -11,6 +11,7 @@ import java.util.List;
 
 @Repository
 public interface ProgramTrainingResourceRepository extends JpaRepository<ProgramTrainingResource, Integer> {
+
     @Query("SELECT ptr FROM ProgramTrainingResource ptr WHERE ptr.educationalResource.resourceId = :resourceId AND ptr.trainingProgram.programId = :programId")
     ProgramTrainingResource findByResourceIdAndProgramId(@Param("resourceId") int resourceId, @Param("programId") int programId);
